Moving the stock class in a separate file.
The docstring are updated to reflect the changes
